+++ internal/mapsx/maps.go
@@ -9,12 +9,3 @@ func Keys[M ~map[K]V, K comparable, V any](m M) []K {
 	}
 	return keys
 }
-
-// Values returns the values of the map m in unspecified order.
-func Values[M ~map[K]V, K comparable, V any](m M) []V {
-	values := make([]V, 0, len(m))
-	for _, v := range m {
-		values = append(values, v)
-	}
-	return values
-}